web_scraping/web_scraping/requirements.txt

botcity-maestro-sdk>=0.3.3,<1.0
botcity-crawler-plugin
# Uncomment to use botcity-utils
# botcity-utils>=0.1.2

# Uncomment to use botcity-crawler-plugin
# botcity-crawler-plugin

# Uncomment to use Pandas
pandas
